requirement.txt
absl-py==1.2.0
args==0.1.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
boto3==1.24.59
botocore==1.27.59
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
clint==0.5.1
comtypes==1.1.14
constantly==15.1.0
cryptography==37.0.4
cycler==0.11.0
daphne==3.0.2
decorator==5.1.1
Django==4.1
dlib==19.24.0
docopt==0.6.2
ecapture==2.0.2
executing==0.10.0
face-recognition==1.3.0
face-recognition-models==0.3.0
feedparser==6.0.10
flatbuffers==1.12
fonttools==4.36.0
future==0.18.2
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gTTS==2.2.4
h5py==3.7.0
hyperlink==21.0.0
idna==3.3
imageio==2.21.1
incremental==21.3.0
ipython==8.4.0
jaraco.context==4.1.2
jedi==0.18.1
jmespath==1.0.1
joblib==1.1.0
Js2Py==0.71
keras==2.9.0
keras-ocr==0.8.9
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
more-itertools==8.14.0
networkx==2.8.5
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipwin==0.5.2
playsound==1.3.0
prompt-toolkit==3.0.30
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycairo==1.21.0
pycparser==2.21
Pygments==2.13.0
PyGObject==3.42.2
pyjokes==0.6.0
pyjsparser==2.7.1
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyPrind==2.11.3
pySmartDL==1.3.4
pytesseract==0.3.10
python-dateutil==2.8.2
python-espeak==0.6.3
python-vlc==3.0.16120
pyttsx3==2.90
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.0
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
stack-data==0.4.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
tk==0.1.0
tkintertable==1.3.3
traitlets==5.3.0
twilio==7.12.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
wikipedia==1.4.0
wolframalpha==5.0.0
wrapt==1.14.1
xmltodict==0.13.0
zope.interface==5.4.0
